✨Feat: signup page -> widgets 폴더로 이동
추가적으로 스탭을 상태관리로 변경되면서 라우팅 수정
